hardware-autotests/mapsme/requirements.txt
pytest <= 4.1.1
attrs <= 19.1.0
pytest-html
requests
Appium-Python-Client<=0.36